Self-Distillation/requirements.txt
datasets==4.3.0
torch==2.9.0
transformers==4.57.1
accelerate==1.11.0
peft==0.17.1
rich==14.2.0
vllm==0.12.0
trl==0.24.0
openai==2.6.1
tqdm==4.67.1
numpy==2.2.6
scipy==1.15.3
matplotlib==3.10.7
flashinfer-python==0.5.3
wandb==0.22.2
pandas==2.3.3
deepspeed==0.18.4
hydra-core>=1.3.0
omegaconf>=2.3.0